requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pymysql
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-dotenv 